+++ electron/src/electron-main.js
@@ -77,7 +77,8 @@ function startSkycoin() {
     '-enable-seed-api=true',
     '-enable-wallet-api=true',
     '-rpc-interface=false',
-    "-disable-csrf=false"
+    '-disable-csrf=false',
+    '-reset-corrupt-db=true'
     // will break
     // broken (automatically generated certs do not work):
     // '-web-interface-https=true',